etl/requirements.txt
elasticsearch==8.8.2
psycopg2==2.9.6
pydantic==2.1.1
python-dotenv==1.0.0
redis==4.6.0
